Fix pointer type

+++ include/bodyprog/bodyprog.h
@@ -836,7 +836,7 @@ extern s32 D_800BCDB0; // Type assumed.
 
 extern u8 D_800BCDD4;
 
-extern s32* D_800BCDE0;
+extern s8* D_800BCDE0; // Type assumed.
 
 extern u16 D_800BCE14;
